+++ viewer/packages/tools/src/AxisGizmo/AxisGizmo.ts
